requirements.txt
streamlit
openai
python-dotenv
streamlit-code-editor
